Pasta frontend

+++ backend/src/main/java/com/dhbrasil/projetoIntegrador/AlugaVerso/service/validation/UserUpdateValidator.java
@@ -29,7 +29,7 @@ public class UserUpdateValidator implements ConstraintValidator<UserUpdateValid,
     public boolean isValid(UserUpdateDTO dto, ConstraintValidatorContext context) {
 
         var uriVars = (Map<String, String>) request.getAttribute(HandlerMapping.URI_TEMPLATE_VARIABLES_ATTRIBUTE);
-        long userId = Long.parseLong(uriVars.get("id"));
+        long userId = Integer.parseInt(uriVars.get("id"));
 
 
         List<FieldMessage> list = new ArrayList<>();
